fixed redis mocks
