+++ test/app.hooks.test.js
@@ -1,17 +1,9 @@
-const chai = require('chai');
+const { expect } = require('chai');
 const { ObjectId } = require('mongoose').Types;
-const mockery = require('mockery');
-const commons = require('@schul-cloud/commons');
-
-const { Configuration } = commons; // separated from require, mocked in tests
-
-const { expect } = chai;
-
 const app = require('../src/app');
 const { sanitizeDataHook } = require('../src/app.hooks');
 const { sanitizeHtml: { sanitizeDeep } } = require('../src/utils');
-const { cleanup, createTestUser, generateRequestParamsFromUser } = require('./services/helpers/testObjects')(app);
-const redisMock = require('./utils/redis/redisMock');
+const { cleanup, createTestUser } = require('./services/helpers/testObjects')(app);
 
 describe('Sanitization Hook', () => {
 	// TODO: Test if it work for create, post and update
@@ -31,7 +23,7 @@ describe('Sanitization Hook', () => {
 			schoolId: '0000d186816abba584714c5f',
 			title: '<script>alert("test");</script>SanitizationTest äöüß§$%/()=',
 			content: '<p>SanitizationTest<script>alert("test);</script>'
-					+ '<a href="javascript:test();">SanitizationTest</a></p>äöüß§$%/()=',
+				+ '<a href="javascript:test();">SanitizationTest</a></p>äöüß§$%/()=',
 		};
 
 		const path = 'news';
@@ -61,9 +53,9 @@ describe('Sanitization Hook', () => {
 			subject: '<script>alert("test");</script>SanitizationTest äöüß§$%/()=',
 			type: 'problem',
 			currentState: '<p>SanitizationTest<script>alert("test);</script>'
-						+ '<a href="javascript:test();">SanitizationTest</a></p>äöüß§$%/()=',
+				+ '<a href="javascript:test();">SanitizationTest</a></p>äöüß§$%/()=',
 			targetState: '<p>SanitizationTest<script>alert("test);</script>'
-						+ '<a href="javascript:test();">SanitizationTest</a></p>äöüß§$%/()=',
+				+ '<a href="javascript:test();">SanitizationTest</a></p>äöüß§$%/()=',
 			category: 'dashboard',
 			schoolId: '0000d186816abba584714c5f',
 		};
@@ -94,7 +86,7 @@ describe('Sanitization Hook', () => {
 		const data = {
 			name: '<script>alert("test");</script>SanitizationTest äöüß§$%/()=',
 			description: '<p>SanitizationTest<script>alert("test);</script>'
-						+ '<a href="javascript:test();">SanitizationTest</a></p>äöüß§$%/()=',
+				+ '<a href="javascript:test();">SanitizationTest</a></p>äöüß§$%/()=',
 			color: '#d32f22',
 			teacherIds: ['0000d213816abba584714c0a'],
 			schoolId: '0000d186816abba584714c5f',
@@ -199,83 +191,5 @@ describe('removeObjectIdInData hook', () => {
 			roles: [],
 		});
 		expect(_id.toString()).to.not.equal(response._id.toString());
-	});
-});
-
-describe('handleAutoLogout hook', () => {
-	let fut;
-	let redisHelper;
-	let configBefore;
-
-
-	before(async () => {
-		configBefore = Configuration.toObject(); // deep copy current config
-		Configuration.set('REDIS_URI', '//validHost:6379');
-		Configuration.set('JWT_TIMEOUT_SECONDS', 7200);
-
-		mockery.enable({
-			warnOnReplace: false,
-			warnOnUnregistered: false,
-			useCleanCache: true,
-		});
-		mockery.registerMock('redis', redisMock);
-		mockery.registerMock('@schul-cloud/commons', commons);
-
-		delete require.cache[require.resolve('../src/utils/redis')];
-		/* eslint-disable global-require */
-		redisHelper = require('../src/utils/redis');
-		fut = require('../src/app.hooks').handleAutoLogout;
-		/* eslint-enable global-require */
-		redisHelper.initializeRedisClient();
-	});
-
-	after(async () => {
-		Configuration.parse(configBefore); // reset config to before state
-		mockery.deregisterAll();
-		mockery.disable();
-		await cleanup();
-	});
-
-	it('whitelisted JWT is accepted and extended', async () => {
-		const user = await createTestUser();
-		const params = await generateRequestParamsFromUser(user);
-		const redisIdentifier = redisHelper.getRedisIdentifier(params.authentication.accessToken);
-		await redisHelper.redisSetAsync(redisIdentifier, 'value', 'EX', 1000);
-		const result = await fut({ params });
-		expect(result).to.not.equal(undefined);
-		const ttl = await redisHelper.redisTtlAsync(redisIdentifier);
-		expect(ttl).to.be.greaterThan(7000);
-	});
-
-	it('not whitelisted JWT is rejected', async () => {
-		const user = await createTestUser();
-		const params = await generateRequestParamsFromUser(user);
-		const redisIdentifier = redisHelper.getRedisIdentifier(params.authentication.accessToken);
-		await redisHelper.redisDelAsync(redisIdentifier);
-		try {
-			await fut({
-				params, app: { Config: { data: { REDIS_URI: '//validHost:6379', JWT_TIMEOUT_SECONDS: 7200 } } },
-			});
-			throw new Error('should have failed');
-		} catch (err) {
-			expect(err.message).to.not.equal('should have failed');
-			expect(err.code).to.equal(401);
-			expect(err.message).to.equal('Session was expired due to inactivity - autologout.');
-		}
-	});
-
-	it('JWT_WHITELIST_ACCEPT_ALL can be set to not auto-logout users', async () => {
-		const user = await createTestUser();
-		const params = await generateRequestParamsFromUser(user);
-		const redisIdentifier = redisHelper.getRedisIdentifier(params.authentication.accessToken);
-		await redisHelper.redisDelAsync(redisIdentifier);
-		const result = await fut({ params });
-		expect(result).to.have.property('params');
-		expect(result).to.have.property('app');
-	});
-
-	it('passes through requests without authorisation', async () => {
-		const response = await fut({ params: {} });
-		expect(response).to.not.eq(undefined);
 	});
 });
